+++ src/vue/pages/BusinessesMy.vue
@@ -1,14 +1,14 @@
 <template>
-  <div class="businesses-all">
+  <div class="businesses-my">
     <template v-if="list.length">
-      <div class="businesses-all__list">
+      <div class="businesses-my__list">
         <div
-          class="businesses-all__list-item-wrp"
+          class="businesses-my__list-item-wrp"
           v-for="item in list"
           :key="item.accountId"
         >
           <button
-            class="businesses-all__list-item-btn"
+            class="businesses-my__list-item-btn"
             @click="selectItem(item)"
           >
             <business-card :business="item" />
@@ -18,9 +18,9 @@
     </template>
 
     <template v-else-if="!list.length && isLoading">
-      <div class="businesses-all__list">
+      <div class="businesses-my__list">
         <div
-          class="businesses-all__list-item-wrp"
+          class="businesses-my__list-item-wrp"
           v-for="item in 5"
           :key="item"
         >
@@ -31,27 +31,16 @@
 
     <template v-else-if="!list.length && !isLoading">
       <no-data-message
-        class="businesses-all__no-data-message"
+        class="businesses-my__no-data-message"
         icon-name="domain"
-        :title="'businesses-all.no-list-title' | globalize"
-        :message="'businesses-all.no-list-msg' | globalize"
+        :title="'businesses-my.no-list-title' | globalize"
+        :message="'businesses-my.no-list-msg' | globalize"
       />
     </template>
 
-    <drawer :is-shown.sync="isDrawerShown">
-      <template slot="heading">
-        {{ 'businesses-all.business-details-title' | globalize }}
-      </template>
-
-      <business-viewer
-        :business="currentBusiness"
-        @business-added="closeDrawerAndUpdateList"
-      />
-    </drawer>
-
-    <div class="businesses-all__requests-collection-loader">
+    <div class="businesses-my__requests-collection-loader">
       <collection-loader
-        class="businesses-all__loader"
+        class="businesses-my__loader"
         :first-page-loader="getList"
         @first-page-load="setList"
         @next-page-load="concatList"
@@ -62,41 +51,34 @@
 </template>
 
 <script>
-import CollectionLoader from '@/vue/common/CollectionLoader'
-import NoDataMessage from '@/vue/common/NoDataMessage'
-import BusinessCard from './businesses-all/BusinessCard'
-import BusinessCardSkeleton from './businesses-all/BusinessCardSkeleton'
-import Drawer from '@/vue/common/Drawer'
-import BusinessViewer from './businesses-all/BusinessViewer'
-
 import { vuexTypes } from '@/vuex'
 import { mapGetters } from 'vuex'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 import { Bus } from '@/js/helpers/event-bus'
 import { api } from '@/api'
+import CollectionLoader from '@/vue/common/CollectionLoader'
+import NoDataMessage from '@/vue/common/NoDataMessage'
+
+import BusinessCard from './businesses-all/BusinessCard'
+import BusinessCardSkeleton from './businesses-all/BusinessCardSkeleton'
 
 import { BusinessRecord } from '@/js/records/entities/business.record'
 import { vueRoutes } from '@/vue-router/routes'
 
 export default {
-  name: 'businesses-all',
+  name: 'businesses-my',
 
   components: {
     CollectionLoader,
     BusinessCard,
     BusinessCardSkeleton,
     NoDataMessage,
-    Drawer,
-    BusinessViewer,
   },
 
   data () {
     return {
       isLoading: false,
       list: [],
-      myBusiness: [],
-      isDrawerShown: false,
-      currentBusiness: {},
     }
   },
 
@@ -106,17 +88,13 @@ export default {
     }),
   },
 
-  async created () {
-    await this.getMyBusiness()
-  },
-
   methods: {
     async getList () {
       this.isLoading = true
 
       let result
       try {
-        const endpoint = `/integrations/dns/businesses`
+        const endpoint = `/integrations/dns/clients/${this.accountId}/businesses`
         result = await api.getWithSignature(endpoint)
       } catch (error) {
         ErrorHandler.processWithoutFeedback(error)
@@ -124,22 +102,6 @@ export default {
 
       this.isLoading = false
       return result
-    },
-
-    async getMyBusiness () {
-      try {
-        const endpoint = `/integrations/dns/clients/${this.accountId}/businesses`
-        const { data } = await api.getWithSignature(endpoint)
-        this.myBusiness = data.map(i => new BusinessRecord(i))
-      } catch (error) {
-        ErrorHandler.processWithoutFeedback(error)
-      }
-    },
-
-    checkIsMyBusiness (currentBusiness) {
-      return Boolean(this.myBusiness.find(business => {
-        return business.id === currentBusiness.id
-      }))
     },
 
     setList (newList) {
@@ -153,26 +115,10 @@ export default {
     },
 
     selectItem (item) {
-      const isMyBusiness = this.checkIsMyBusiness(item)
-      if (isMyBusiness) {
-        Bus.emit('businesses:setCurrentBusiness', {
-          business: item,
-          redirectTo: vueRoutes.movements,
-        })
-      } else {
-        this.currentBusiness = item
-        this.isDrawerShown = true
-      }
-    },
-
-    reloadList () {
-      return this.$refs.listCollectionLoader.loadFirstPage()
-    },
-
-    async closeDrawerAndUpdateList () {
-      this.isDrawerShown = false
-      this.reloadList()
-      await this.getMyBusiness()
+      Bus.emit('businesses:setCurrentBusiness', {
+        business: item,
+        redirectTo: vueRoutes.movements,
+      })
     },
   },
 }
@@ -185,30 +131,30 @@ export default {
 $list-item-margin: 2rem;
 $filter-field-to-filter-field-margin: 2rem;
 
-.businesses-all__filters {
+.businesses-my__filters {
   margin: -$filter-field-to-filter-field-margin 0 2.4rem
     (-$filter-field-to-filter-field-margin);
 }
 
-.businesses-all__filter-field {
+.businesses-my__filter-field {
   margin: $filter-field-to-filter-field-margin 0 0
     $filter-field-to-filter-field-margin;
 }
 
-.businesses-all__list-item-btn {
+.businesses-my__list-item-btn {
   display: block;
   width: 100%;
   max-width: 100%;
   text-align: left;
 }
 
-.businesses-all__list {
+.businesses-my__list {
   display: flex;
   flex-wrap: wrap;
   margin: -$list-item-margin 0 0 (-$list-item-margin);
 }
 
-.businesses-all__list-item-wrp {
+.businesses-my__list-item-wrp {
   margin: $list-item-margin 0 0 $list-item-margin;
   width: calc(100% + #{$list-item-margin});
 
@@ -238,11 +184,11 @@ $filter-field-to-filter-field-margin: 2rem;
   }
 }
 
-.businesses-all__loader {
+.businesses-my__loader {
   margin-top: 1rem;
 }
 
-.businesses-all__no-data-message {
+.businesses-my__no-data-message {
   margin-top: 4.8rem;
 }
 </style>